feat: table body row as separate component
