+++ components/src/components/table/table-body-row/readme.md
@@ -1,0 +1,32 @@
+# sdds-table-body-row
+
+
+
+<!-- Auto Generated Below -->
+
+
+## Events
+
+| Event            | Description                                                                                                             | Type                   |
+| ---------------- | ----------------------------------------------------------------------------------------------------------------------- | ---------------------- |
+| `bodyRowToTable` | Send status of single row to the parent, sdds-table component that hold logic for data export and main checkbox control | `CustomEvent<boolean>` |
+
+
+## Dependencies
+
+### Used by
+
+ - [sdds-table](..)
+ - [sdds-table-body-row-expendable](..)
+
+### Graph
+```mermaid
+graph TD;
+  sdds-table --> sdds-table-body-row
+  sdds-table-body-row-expendable --> sdds-table-body-row
+  style sdds-table-body-row fill:#f9f,stroke:#333,stroke-width:4px
+```
+
+----------------------------------------------
+
+*Built with [StencilJS](https://stenciljs.com/)*
